requirements.txt
opencv-python
opencv-contrib-python
numpy
scikit-learn
scikit-image
imutils
Pillow
matplotlib
seaborn
joblib
